+++ SoundFontInfoLib/SoundFontInfoLib/Entity/SampleHeader.hpp
@@ -81,7 +81,7 @@ public:
   /// @returns the pitch correction to apply when playing back the samples
   short pitchCorrection() const { return correction; }
   
-  void dump(const std::string& indent, int index) const;
+  void dump(const std::string& indent, size_t index) const;
   
 private:
   std::string sampleTypeDescription() const;
